Refactor image processing and NMS functions for improved clarity and performance
This commit enhances the `invert` function by directly using `std::numeric_limits` within the transformation lambda, improving readability. Additionally, the `nms` function is updated to use `static_cast<size_t>` for index access, ensuring type safety. The `BoundingBox` constructor parameters are renamed for better clarity, enhancing code maintainability.

+++ nextcv/_cpp/src/postprocessing/nms.cpp
@@ -17,26 +17,27 @@ std::vector<int> nms(const std::vector<std::array<float, 4>>& bboxes,
     std::vector<int> indices(bboxes.size());
     std::iota(indices.begin(), indices.end(), 0);
 
-    std::sort(indices.begin(), indices.end(),
-              [&scores](int a, int b) { return scores[a] > scores[b]; });
+    std::sort(indices.begin(), indices.end(), [&scores](int a, int b) {
+        return scores[static_cast<size_t>(a)] > scores[static_cast<size_t>(b)];
+    });
 
     std::vector<int> result;
     std::vector<bool> suppressed(bboxes.size(), false);
 
     for (size_t i = 0; i < indices.size(); ++i) {
-        if (suppressed[indices[i]])
+        if (suppressed[static_cast<size_t>(indices[i])])
             continue;
 
         result.push_back(indices[i]);
 
         // Suppress boxes with high IoU
         for (size_t j = i + 1; j < indices.size(); ++j) {
-            if (suppressed[indices[j]])
+            if (suppressed[static_cast<size_t>(indices[j])])
                 continue;
 
             // Calculate IoU
-            const auto& box1 = bboxes[indices[i]];
-            const auto& box2 = bboxes[indices[j]];
+            const auto& box1 = bboxes[static_cast<size_t>(indices[i])];
+            const auto& box2 = bboxes[static_cast<size_t>(indices[j])];
 
             // Intersection
             float x1 = std::max(box1[0], box2[0]);
@@ -54,7 +55,7 @@ std::vector<int> nms(const std::vector<std::array<float, 4>>& bboxes,
             float iou = intersection / union_area;
 
             if (iou > threshold) {
-                suppressed[indices[j]] = true;
+                suppressed[static_cast<size_t>(indices[j])] = true;
             }
         }
     }
